Well a picture displays...
I've got a picture displaying. It's hardwired in and is not
formatted at all. But it's a step in the right direction. What
makes my life slightly more complicated now is that I need to have
recipes include a picture resource. Shouldn't be hard, but it is
just anohter thing to work on. I'll tend to that later.