Some major improvements tonight.
So I moved everything into seperate files. Mostly because fk.kv was
getting a little bloated. This will make it easier to keep track of.
but I've still got a fair bit of structual weird-ness I think. Also
no decoupling of model or view. Things I will work on in the future.

I made use of my new app Kivy Kazam to make the interface for the
meal plan editor. It's nice having a real time updating interface
drafting system.

I was having issues with connecting the meal plan name on the plist
to the name being entered in the editor. I realized at the end of
the evening that I should do this through properties. Some of the
tricks and methods of doing things when i was programming XAML are
coming back to me. I dread to call myself a interface programmer,
and I still don't believe I am. But I'm getting better. Or at least
good enough.

I also made some updates for the android app. It can now go portait
or landscape, and it should crash in the process. You can leave and
come back with out having to re-load the program. And it should say
it is a product of Studio Sleep Giraffe!